+++ opengl_renderer/src/Graphics/MeshFactory.cpp
@@ -7,19 +7,19 @@
 
 namespace Real {
 
-    std::pair<std::vector<Graphics::Vertex>, std::vector<uint32_t>> MeshFactory::CreateTriangle() {
+    std::pair<std::vector<Graphics::Vertex>, std::vector<uint64_t>> MeshFactory::CreateTriangle() {
         std::vector<Graphics::Vertex> vertices = {
             { glm::vec3(-0.5f, -0.5f, 0.0f), glm::vec3(0.0f, 0.0f, 1.0f), glm::vec2(0.0f, 0.0f) },
             { glm::vec3( 0.5f, -0.5f, 0.0f), glm::vec3(0.0f, 0.0f, 1.0f), glm::vec2(1.0f, 0.0f) },
             { glm::vec3( 0.0f,  0.5f, 0.0f), glm::vec3(0.0f, 0.0f, 1.0f), glm::vec2(0.5f, 1.0f) }
         };
 
-        std::vector<uint32_t> indices = { 0, 1, 2 };
+        std::vector<uint64_t> indices = { 0, 1, 2 };
 
         return std::make_pair(vertices, indices);
     }
 
-    std::pair<std::vector<Graphics::Vertex>, std::vector<uint32_t>> MeshFactory::CreateCube() {
+    std::pair<std::vector<Graphics::Vertex>, std::vector<uint64_t>> MeshFactory::CreateCube() {
         std::vector<Graphics::Vertex> vertices =
 		{	// Vertices, Normal, UV
 
@@ -60,7 +60,7 @@ namespace Real {
 			{ glm::vec3(-0.5f, -0.5f,  0.5f),  glm::vec3 {0.0, -1.0, 0.0},  {0.0f, 1.0f}  },
 		};
 
-		std::vector<uint32_t> indices =
+		std::vector<uint64_t> indices =
 		{
 			// Front face
 			0, 1, 2,      2, 3, 0,
